RestApi/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-annoying==0.10.6
django-jsonfield==1.4.1
djangorestframework==3.12.4
idna==2.10
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
Werkzeug==1.0.1
jsonfield==3.1.0
itsdangerous==1.1.0
django-crispy-forms==1.11.2